fix(int16): expand bias_q from int32 to int64 to fix overflow
For INT16 quantization, bias_q = bias / (scale_in * scale_w) can reach
~1e10, exceeding int32 max (2.1e9). This caused severe accuracy loss
(cosine 0.118 vs expected ~0.999).

Root cause: scale_in (~1.5e-5) * scale_w (~3e-6) = ~4.5e-11, making
the denominator extremely small for INT16. Typical bias values (~0.3)
divided by this produce values far exceeding int32 range.

Changes:
- perchannel_param_t.bias_q: int32_t -> int64_t (10 -> 14 bytes/channel)
- Accumulator uses int64 throughout (conv, dwconv, fc, pooling, etc.)
- Add residual_src field for skip connections
- Add DUMP_LAYERS debug feature
- Add E2E test for eltwise add

Results after fix:
- INT8 per-channel:  cosine = 0.974 (unchanged)
- INT16 per-channel: cosine = 0.118 -> 0.99999

+++ include/npu_operators.h
@@ -13,7 +13,8 @@
 /*
  * All operators:
  *   - Input/output tensors in NHWC layout
- *   - Return raw INT32 accumulator results (before post-processing)
+ *   - Return raw INT64 accumulator results (before post-processing)
+ *     Hardware uses 40-bit accumulator; we model with 64-bit for correctness.
  *   - Weights in appropriate layout per operator
  *   - bias array: INT32, one per output channel
  */
@@ -25,7 +26,7 @@ void npu_conv2d(const layer_config_t *cfg,
                 const tensor_t *input,
                 const int8_t *weights,
                 const int32_t *bias,
-                int32_t *output_acc);  /* [out_h * out_w * out_c] */
+                int64_t *output_acc);  /* [out_h * out_w * out_c] */
 
 /* Depthwise Conv: per-channel convolution
  * weights layout: [channels][kernel_h][kernel_w]
@@ -34,7 +35,7 @@ void npu_dwconv(const layer_config_t *cfg,
                 const tensor_t *input,
                 const int8_t *weights,
                 const int32_t *bias,
-                int32_t *output_acc);
+                int64_t *output_acc);
 
 /* Fully Connected: equivalent to 1x1 conv with h=w=1
  * weights layout: [out_c][in_c]
@@ -43,14 +44,14 @@ void npu_fc(const layer_config_t *cfg,
             const tensor_t *input,
             const int8_t *weights,
             const int32_t *bias,
-            int32_t *output_acc);
+            int64_t *output_acc);
 
 /* Pooling: Max or Average
- * No weights needed. Output is INT32 (for Avg, sum before divide).
+ * No weights needed. Output is INT64 accumulator.
  */
 void npu_pooling(const layer_config_t *cfg,
                  const tensor_t *input,
-                 int32_t *output_acc);
+                 int64_t *output_acc);
 
 /* Eltwise Add: element-wise addition of two tensors
  * Both inputs must have same dimensions.
@@ -58,12 +59,12 @@ void npu_pooling(const layer_config_t *cfg,
 void npu_eltwise_add(const layer_config_t *cfg,
                      const tensor_t *input_a,
                      const tensor_t *input_b,
-                     int32_t *output_acc);
+                     int64_t *output_acc);
 
 /* Resize: nearest-neighbor or bilinear upsampling/downsampling */
 void npu_resize(const layer_config_t *cfg,
                 const tensor_t *input,
-                int32_t *output_acc);
+                int64_t *output_acc);
 
 /* Deconv: transposed convolution (insert zeros then conv)
  * weights layout: same as conv2d [out_c][kernel_h][kernel_w][in_c]
@@ -72,7 +73,7 @@ void npu_deconv(const layer_config_t *cfg,
                 const tensor_t *input,
                 const int8_t *weights,
                 const int32_t *bias,
-                int32_t *output_acc);
+                int64_t *output_acc);
 
 /* Concat: channel-wise concatenation
  * Copies input into output at the specified channel offset.
